DDF-4320 Move and rename some packages
