+++ catalog/transformer/catalog-transformer-bootflag/src/test/resources/valid-transformers/notajsonfile.txt
