+++ distribution/ddf-common/src/main/resources/etc/transformers/README.md
@@ -9,7 +9,7 @@ To allow the CDM to wait for additional Input Transformers, drop a new JSON file
 To increase or decrease the amount of time required to wait for Input Transformers use the following system property. The timeout
 cannot be set to below 30 seconds. The default timeout is set to 5 minutes.
 ```
-org.codice.ddf.platform.osgi.transformerWaitTimeoutSeconds
+org.codice.ddf.platform.bootflag.transformerWaitTimeoutSeconds
 ```
 
 ## File Format
