--- platform/osgi/serviceflag-inputtransformer/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-/**
- * Copyright (c) Codice Foundation
- *
- * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
- * version 3 of the License, or any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
- * <http://www.gnu.org/licenses/lgpl.html>.
- *
- **/
-
- -->
-<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
-           http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd">
-
-    <reference-list id="inputTransformers" interface="ddf.catalog.transform.InputTransformer"
-                    member-type="service-reference" filter="(id=*)" availability="optional"/>
-
-    <bean id="inputTransformerIds" class="org.codice.ddf.platform.serviceflag.inputtransformer.InputTransformerIds"/>
-
-    <bean id="inputTransformerServiceFlag"
-          class="org.codice.ddf.platform.serviceflag.inputtransformer.InputTransformerBootServiceFlag">
-        <argument ref="inputTransformerIds"/>
-        <argument ref="inputTransformers"/>
-    </bean>
-</blueprint>
